refactored and ws ssl tests
